\name{identify}
\alias{identify}
\title{Learn about the Pensoft OAI-PMH service.}
\usage{
  identify(url = "http://oai.pensoft.eu",
    transform = FALSE)
}
\arguments{
  \item{url}{Base URL for the service.}
}
\value{
  List of information describing the Pensoft OAI-PMH
  services.
}
\description{
  Learn about the Pensoft OAI-PMH service.
}
\examples{
\dontrun{
identify()
}
}

