\name{ps_identify}
\alias{ps_identify}
\title{Learn about the Pensoft OAI-PMH service.}
\usage{
  ps_identify(url = "http://oai.pensoft.eu",
    transform = TRUE)
}
\arguments{
  \item{url}{Base URL for the service.}
}
\value{
  List of information describing the Pensoft OAI-PMH
  services.
}
\description{
  Learn about the Pensoft OAI-PMH service.
}
\examples{
\dontrun{
ps_identify()
}
}

